requirements.txt
# transformers>=4.51.3
git+https://github.com/huggingface/transformers.git
accelerate>=1.12.0
librosa
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
ffmpeg-python>=0.2.0
python-multipart>=0.0.6
torchaudio>=2.0.0
pydub>=0.25.1
loguru

openai-whisper